feat: obatalkes create

+++ siFARMASI/src/main/java/apap/TA_B1/siFARMASI/repository/ObatAlkesDb.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface ObatAlkesDb extends JpaRepository<ObatAlkesModel, Integer> {
-    ObatAlkesModel getObatAlkesById(Long id);
+    ObatAlkesModel getObatAlkesById(Integer id);
 }
